fixed codeclimate issues
